separate out the interstitial page
